pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["mac <risephoenixpuryfire@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
aiogram = "^3.12.0"
pydentic = "^0.0.1.dev3"
logging = "^0.4.9.6"
sqlalchemy = "^2.0.32"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
